Moves CacheMode to volume service & Makes use of urllib module
