+++ intern/service/boss/baseversion.py
@@ -15,12 +15,7 @@ import six
 from abc import ABCMeta, abstractmethod
 from intern.resource.boss.resource import CoordinateFrameResource
 from requests import Request
-from enum import Enum
-
-class CacheMode(str, Enum):
-    cache = 'cache'
-    no_cache = 'no-cache'
-    raw = 'raw'
+import urllib
 
 @six.add_metaclass(ABCMeta)
 class BaseVersion(object):
@@ -152,7 +147,7 @@ class BaseVersion(object):
         return urlWithKey + '&value=' + str(value)
 
     def build_cutout_url(
-        self, resource, url_prefix, resolution, x_range, y_range, z_range,  access_mode=CacheMode.no_cache, time_range=None, id_list=[]):
+        self, resource, url_prefix, resolution, x_range, y_range, z_range, time_range=None, id_list=[], access_mode=None):
         """Build the url to access the cutout function of the Boss' volume service.
 
         Args:
@@ -167,6 +162,7 @@ class BaseVersion(object):
                 cache = Will check both cache and for dirty keys
                 no_cache = Will skip cache check but check for dirty keys
                 raw = Will skip both the cache and dirty keys check
+                NOTE: No default here since when building a cutout url to create_cutout there is no need for access_mode
 
         Returns:
             (string): Full URL to access API.
@@ -187,12 +183,15 @@ class BaseVersion(object):
             t_rng_lst = self.convert_int_list_range_to_str(time_range)
             urlWithParams += t_rng_lst + '/'
 
+        queryParamDict = {}
         if len(id_list) > 0:
-            urlWithParams += '?filter=' + self.convert_int_list_to_comma_sep_str(id_list)
-
-        # Add the access_mode parameter
-        urlWithParams += '?access-mode={}'.format(access_mode)
-
+            queryParamDict['filter'] = self.convert_int_list_to_comma_sep_str(id_list)
+    
+        # If creating a cutout, the url will not include the access_mode otherwise it will
+        if access_mode is not None:
+            queryParamDict['access-mode'] = access_mode.value 
+        if queryParamDict:
+            urlWithParams += '?' + urllib.parse.urlencode(queryParamDict,safe=",")
         return urlWithParams
 
     def get_request(self, resource, method, content, url_prefix, token, proj_list_req=False, json=None, data=None):
@@ -262,7 +261,7 @@ class BaseVersion(object):
 
     def get_cutout_request(
         self, resource, method, content, url_prefix, token,
-        resolution, x_range, y_range, z_range, time_range, access_mode=CacheMode.no_cache, numpyVolume=None, id_list=[]):
+        resolution, x_range, y_range, z_range, time_range,  numpyVolume=None, id_list=[], access_mode=None,):
 
         """Create a request for working with cutouts (part of the Boss' volume service).
 
@@ -291,7 +290,7 @@ class BaseVersion(object):
             RuntimeError if url_prefix is None or an empty string.
         """
         url = self.build_cutout_url(
-            resource, url_prefix, resolution, x_range, y_range, z_range, access_mode, time_range, id_list)
+            resource, url_prefix, resolution, x_range, y_range, z_range,  time_range, id_list, access_mode)
         headers = self.get_headers(content, token)
         return Request(method, url, headers = headers, data = numpyVolume)
